refactor(library/ShimmedStabilityAIClient): extracts "toShortfinPipelineInputText" module
